Mutualize CommitterRepository
