+++ config/packages/framework.yaml
@@ -29,7 +29,7 @@ framework:
     webhook:
         routing:
             github:
-                service: App\Shared\Webhook\GithubWebhookParser
+                service: App\Shared\Infrastructure\Webhook\GithubWebhookParser
 
 when@test:
     framework:
