+++ config/services.php
@@ -1,15 +1,13 @@
 <?php
 
-use App\PullRequest\Domain\Gateway\CommitterRepositoryInterface as PRCommitterRepositoryInterface;
 use App\PullRequest\Domain\Gateway\PullRequestRepositoryInterface;
-use App\PullRequest\Infrastructure\Adapter\RestGithubCommitterRepository as PRRestGithubCommitterRepository;
 use App\PullRequest\Infrastructure\Adapter\RestPullRequestRepository;
-use App\PullRequestDashboard\Domain\Gateway\CommitterRepositoryInterface;
 use App\PullRequestDashboard\Domain\Gateway\PullRequestCardRepositoryInterface;
 use App\PullRequestDashboard\Infrastructure\Adapter\GraphqlGithubPullRequestCardRepository;
-use App\PullRequestDashboard\Infrastructure\Adapter\RestGithubCommitterRepository;
-use App\Shared\Adapter\SpyMessageBus;
-use App\Shared\Factory\CommandFactory\CommandFactory;
+use App\Shared\Domain\Gateway\CommitterRepositoryInterface;
+use App\Shared\Infrastructure\Adapter\RestGithubCommitterRepository;
+use App\Shared\Infrastructure\Adapter\SpyMessageBus;
+use App\Shared\Infrastructure\Factory\CommandFactory\CommandFactory;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 use Symfony\Component\Messenger\MessageBusInterface;
 
@@ -44,7 +42,6 @@ return function (ContainerConfigurator $configurator) {
     $services->alias(PullRequestRepositoryInterface::class, RestPullRequestRepository::class);
     $services->alias(PullRequestCardRepositoryInterface::class, GraphqlGithubPullRequestCardRepository::class);
     $services->alias(CommitterRepositoryInterface::class, RestGithubCommitterRepository::class);
-    $services->alias(PRCommitterRepositoryInterface::class, PRRestGithubCommitterRepository::class);
     $services->set(CommandFactory::class)
         ->args([tagged_iterator('app.shared.command_strategy')])
     ;
@@ -66,7 +63,6 @@ return function (ContainerConfigurator $configurator) {
         $services->set(RestPullRequestRepository::class);
         $services->set(GraphqlGithubPullRequestCardRepository::class);
         $services->set(RestGithubCommitterRepository::class);
-        $services->set(PRRestGithubCommitterRepository::class);
         $services->alias(MessageBusInterface::class, SpyMessageBus::class);
     }
 };
