--- src/Shared/Event/GithubEvent.php
@@ -1,14 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Shared\Event;
-
-class GithubEvent
-{
-    public function __construct(
-        public readonly string $eventType,
-        public readonly string $payload,
-    ) {
-    }
-}
