+++ src/Shared/Infrastructure/Event/GithubEvent.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Shared\Infrastructure\Event;
+
+class GithubEvent
+{
+    public function __construct(
+        public readonly string $eventType,
+        public readonly string $payload,
+    ) {
+    }
+}
